Application Layout hook renamed & Sidenav updated
