+++ lib/src/sidenav/Sidenav.tsx
@@ -1,41 +1,54 @@
-// @ts-nocheck
 import React from "react";
 import styled, { ThemeProvider } from "styled-components";
+import { spaces, responsiveSizes } from "../common/variables.js";
 
-import { spaces } from "../common/variables.js";
 import useTheme from "../useTheme";
 import { BackgroundColorProvider } from "../BackgroundColorContext";
-import SidenavPropsType, { SidenavTitlePropsType, SidenavSubtitlePropsType, SidenavLinkPropsType } from "./types.js";
+import SidenavPropsType, { SidenavTitlePropsType, SidenavSubtitlePropsType, SidenavLinkPropsType } from "./types";
+import { useResponsiveSidenavVisibility } from "../layout/SidenavContext";
 
 const DxcSidenav = ({ padding, children }: SidenavPropsType): JSX.Element => {
   const colorsTheme = useTheme();
 
   return (
     <ThemeProvider theme={colorsTheme.sidenav}>
-      <SideNavContainer padding={padding}>
+      <SidenavContainer padding={padding}>
         <BackgroundColorProvider color={colorsTheme.sidenav.backgroundColor}>{children}</BackgroundColorProvider>
-      </SideNavContainer>
+      </SidenavContainer>
     </ThemeProvider>
   );
 };
 
-const Title = ({ children }: SidenavTitlePropsType): JSX.Element => <SideNavMenuTitle>{children}</SideNavMenuTitle>;
+const Title = ({ children }: SidenavTitlePropsType): JSX.Element => <SidenavMenuTitle>{children}</SidenavMenuTitle>;
 
 const Subtitle = ({ children }: SidenavSubtitlePropsType): JSX.Element => (
-  <SideNavMenuSubTitle>{children}</SideNavMenuSubTitle>
+  <SidenavMenuSubTitle>{children}</SidenavMenuSubTitle>
 );
 
-const Link = ({ tabIndex = 0, href, onClick, children }: SidenavLinkPropsType): JSX.Element => (
-  <SideNavMenuLink tabIndex={tabIndex} href={href} onClick={onClick}>
-    {children}
-  </SideNavMenuLink>
-);
+const Link = ({ tabIndex = 0, href, onClick, children }: SidenavLinkPropsType): JSX.Element => {
+  const setIsSidenavVisibleResponsive = useResponsiveSidenavVisibility();
+  const handleClick = () => {
+    onClick?.();
+    setIsSidenavVisibleResponsive?.(false);
+  };
 
-const SideNavContainer = styled.div`
+  return (
+    <SidenavMenuLink tabIndex={tabIndex} href={href} onClick={handleClick}>
+      {children}
+    </SidenavMenuLink>
+  );
+};
+
+const SidenavContainer = styled.div<SidenavPropsType>`
   display: flex;
   flex-direction: column;
   background-color: ${(props) => props.theme.backgroundColor};
-  max-width: 300px;
+
+  @media (max-width: ${responsiveSizes.medium}rem) {
+    width: ${(props) =>
+      props.padding ? `calc(100vw - ${spaces[props.padding]} - ${spaces[props.padding]})` : "100vw"};
+  }
+
   width: ${(props) => (props.padding ? `calc(300px - ${spaces[props.padding]} - ${spaces[props.padding]})` : "300px")};
   padding: ${(props) => (props.padding ? spaces[props.padding] : "")};
 
@@ -54,7 +67,7 @@ const SideNavContainer = styled.div`
   }
 `;
 
-const SideNavMenuTitle = styled.div`
+const SidenavMenuTitle = styled.div`
   font-family: ${(props) => props.theme.titleFontFamily};
   font-size: ${(props) => props.theme.titleFontSize};
   font-style: ${(props) => props.theme.titleFontStyle};
@@ -65,7 +78,7 @@ const SideNavMenuTitle = styled.div`
   margin-bottom: 16px;
 `;
 
-const SideNavMenuSubTitle = styled.div`
+const SidenavMenuSubTitle = styled.div`
   font-family: ${(props) => props.theme.subtitleFontFamily};
   font-size: ${(props) => props.theme.subtitleFontSize};
   font-style: ${(props) => props.theme.subtitleFontStyle};
@@ -76,7 +89,7 @@ const SideNavMenuSubTitle = styled.div`
   margin-bottom: 4px;
 `;
 
-const SideNavMenuLink = styled.a`
+const SidenavMenuLink = styled.a`
   font-family: ${(props) => props.theme.linkFontFamily};
   font-size: ${(props) => props.theme.linkFontSize};
   font-style: ${(props) => props.theme.linkFontStyle};
